Add AnyIO to dependencies
https://anyio.readthedocs.io/en/stable/index.html
https://anyio.readthedocs.io/en/stable/testing.html
https://anyio.readthedocs.io/en/stable/fileio.html
https://github.com/encode/starlette/releases/tag/0.15.0

AnyIO provides tools for asynchronous Python operations.
To start, AnyIO will be used for asyncio support with pytest.
The file I/O features may also be used when loading .env files.
Of note, AnyIO is the first external dependency for Starlette.

AnyIO will be an optional dependency, so the GitHub Actions workflow
will be updated to ensure that it is installed.