Require AnyIO
https://anyio.readthedocs.io/en/stable/index.html
https://anyio.readthedocs.io/en/stable/versionhistory.html
https://github.com/encode/starlette/releases/tag/0.15.0

AnyIO was initially installed as an optional dependency. Late imports
and exception handling were used to allow functions to do file I/O with
AnyIO, while handling `ImportError`s if the package was not installed.

This commit will make AnyIO a required dependency. This is reasonable,
given the current usefulness of the file I/O features, the potential
usefulness of the `anyio.Path` features added in version 3.3.0, and the
fact that AnyIO is the first required external dependency of Starlette.
It also simplifies the logic in the `dotenv` module and tests.

The `all` Poetry extras list will likely be populated with optional
dependencies as the project progresses, so it will be retained.