Pin and test Poetry version
Poetry has introduced a number of problematic changes recently. To help
buffer against these problematic changes, this commit will implement
installation and testing of specific versions of Poetry, rather than the
latest version.

The following steps will be taken to install Poetry and verify that the
correct version is installed:

1. Set the `PIPX_VERSION` and `POETRY_VERSION` environment variables,
   which will be used to install specific versions of each package
2. Install `pipx` with `pip`, for the appropriate version of Python:
   `pipx` is included by default in the GitHub Actions virtual
   environment, but only for the default Python version, not necessarily
   the version installed by actions/setup-python
3. Install Poetry with `pipx`, instead of the get-poetry.py or
   install-poetry.py install scripts
4. Run a shell test to assert that the Poetry version returned by the
   `poetry -V` command matches `POETRY_VERSION`