Add default filename to `load_dotenv` docs
br3ndonland/fastenv#5

The `load_dotenv` method originally had a default filename argument of
`".env"`. This default was removed in #5, when the ability to load
multiple .env files was implemented. The default filename argument could
potentially be added back in the future, but adding the same default to
the examples in the docs will help ensure the examples work, regardless
of whether or not `load_dotenv` supplies a default filename argument.