Address CodeQL URL substring sanitization error
Incomplete URL substring sanitization (High):
The string digitaloceanspaces.com may be at an arbitrary position in
the sanitized URL.

Note that this issue occurs in a test modified by commit aaf3f1d, so
it's not really a security vulnerability. It's an easy fix though. This
commit will resolve the substring sanitization error.