Don't hard-code repo name in GitHub Actions jobs
Commit d0e973c set up dedicated workflow jobs for PyPI by uploading the
Python package as a build artifact in the first job, then downloading
and publishing the package in a separate job. The artifact needs a name.
GitHub Actions offers a `GITHUB_REPOSITORY`/`${{ github.repository }}`
variable, with a value in the format `org-name/repo-name`. This variable
was used in the setup job but not output for use in other jobs.

This commit will output the repo name from the setup job so it can be
used in the other workflow jobs instead of hard-coding the repo name.