Switch type checking from mypy to BasedPyright
This commit will switch Python type checking from mypy to
[BasedPyright](https://docs.basedpyright.com/latest/). See the
[docs](https://docs.basedpyright.com/latest/) for comparisons with mypy
and Pyright.

Pre-existing errors have been fixed if possible. BasedPyright supports
[baselining](https://docs.basedpyright.com/latest/benefits-over-pyright/baseline/),
in which existing errors can be written to a file. This allows adoption
of the type checker without having to fix all pre-existing errors first.
The baseline file will be added to `.basedpyright/baseline.json`.

The rule `reportImplicitOverride` will be set to "information" only. The
fix for this warning is to use the `@typing.override` decorator that was
introduced in Python 3.12. The `typing_extensions` package provides this
decorator for use with older versions of Python is needed. The problem
is that, while the
[typing docs](https://docs.python.org/3/library/typing.html#typing.override)
say "There is no runtime checking of this property," use as a decorator
(rather than just a type annotation) means it is required at runtime.
Therefore, this project would have to add a runtime dependency on the
`typing_extensions` package, which is not of interest at this time.

In addition to superior type checking results, another motivation for
this change is the language server provided by the BasedPyright
VSCode/VSCodium extension.
[Jedi](https://github.com/pappasam/jedi-language-server) is no longer
working well for
[Python editing](https://code.visualstudio.com/docs/python/editing)
features like autocompletion and go to definition, and the Microsoft
Pylance/Pyright extension is only available on the proprietary Microsoft
extension registry. BasedPyright is another option also available on
[Open VSX](https://open-vsx.org/extension/detachhead/basedpyright) for
users of [VSCodium](https://code.visualstudio.com/docs/python/editing).
This commit will add an `extensions.json` file to suggest BasedPyright.

Note that the VSCode `settings.json` for the user should include
`"python.languageServer": "None"` so that BasedPyright can act as the
language server. `"None"` is currently the default for this setting, but
it could change in the future, so it is prudent to explicitly set it to
`"None"`. This setting cannot be added at the folder level.